LayoutTests/imported/w3c/web-platform-tests/mathml/presentation-markup/operators/embellished-operator-001-expected.txt


FAIL mrow (embellished operator) assert_approx_equals: expected 50 +/- 1 but got 0
FAIL mrow (not embellished operator) assert_approx_equals: expected 50 +/- 1 but got 0
FAIL mstyle (embellished operator) assert_approx_equals: expected 50 +/- 1 but got 0
FAIL mstyle (not embellished operator) assert_approx_equals: expected 50 +/- 1 but got 0
FAIL mphantom (embellished operator) assert_approx_equals: expected 50 +/- 1 but got 0
FAIL mphantom (not embellished operator) assert_approx_equals: expected 50 +/- 1 but got 0
FAIL mpadded (embellished operator) assert_approx_equals: expected 50 +/- 1 but got 0
FAIL mpadded (not embellished operator) assert_approx_equals: expected 50 +/- 1 but got 0
X
X
X
X

X
X
X
X

X
X
X

X
X
X
X

X
X

X
X

X
X
X

X
X
X
X
